refactor modal video and set into sample
