--- src/containers/home/stories/samples/modal-video/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
