+++ src/containers/home/stories/samples/sample-composition/sample-media/component.tsx
@@ -1,12 +1,28 @@
 import React, { useEffect, useRef } from 'react';
 
+import dynamic from 'next/dynamic';
+
 import { motion, useScroll, useTransform } from 'framer-motion';
+import useBreakpoint from 'use-breakpoint';
+
+import { useModal } from 'hooks/modals';
+
+import Modal from 'components/modal';
+import PlayControl from 'components/video/play-control';
+import { BREAKPOINTS } from 'styles/styles.config';
 
 import type { SampleMediaProps } from './types';
+const Video = dynamic(() => import('components/video'), {
+  ssr: false,
+});
 
-export const SampleMedia = ({ backgroundImage }: SampleMediaProps) => {
+export const SampleMedia = ({ backgroundImage, video = false }: SampleMediaProps) => {
   const targetRef = useRef(null);
   const containerRef = useRef(null);
+  const { minWidth } = useBreakpoint(BREAKPOINTS, 'md');
+
+  const { isOpen, open, close } = useModal();
+
   const { scrollYProgress } = useScroll({
     container: containerRef,
     target: targetRef,
@@ -15,20 +31,60 @@ export const SampleMedia = ({ backgroundImage }: SampleMediaProps) => {
 
   const bgPos = useTransform(scrollYProgress, [0, 1], [`50% 70%`, `50% 30%`]);
 
+  const videoConfig = {
+    youtube: {
+      playerVars: {
+        controls: true,
+        showinfo: 0,
+        rel: 0,
+      },
+    },
+  };
+
   useEffect(() => {
     const el = document.getElementById('scroll-container');
     containerRef.current = el;
   });
 
   return (
-    <motion.div
-      ref={targetRef}
-      className="overflow-hidden bg-center bg-no-repeat bg-cover aspect-video"
-      style={{
-        backgroundImage,
-        backgroundPosition: bgPos,
-      }}
-    />
+    <div className="relative">
+      <motion.div
+        ref={targetRef}
+        className="overflow-hidden bg-center bg-no-repeat bg-cover aspect-video"
+        style={{
+          backgroundImage,
+          backgroundPosition: bgPos,
+        }}
+      />
+      {video && (
+        <button
+          onClick={() => open()}
+          className="absolute top-[50%] left-[50%] transform -translate-x-1/2 -translate-y-1/2 z-10"
+        >
+          <PlayControl playBackground="bg-white" playColor="text-black" />
+        </button>
+      )}
+      <Modal
+        title=""
+        size="l"
+        open={isOpen}
+        closeBtn={minWidth >= BREAKPOINTS.sm ? false : true}
+        onOpenChange={() => close()}
+      >
+        <div className="w-full translate-y-full sm:translate-y-0 sm:pt-0 aspect-video">
+          <Video
+            config={videoConfig}
+            playing
+            light={false}
+            loop
+            url="https://youtu.be/6PGyRolwFVo"
+            height="100%"
+            width="100%"
+            muted
+          />
+        </div>
+      </Modal>
+    </div>
   );
 };
