+++ src/containers/home/stories/samples/sample-composition/sample-media/types.d.ts
@@ -1,3 +1,4 @@
 export interface SampleMediaProps {
   backgroundImage?: string;
+  video?: boolean;
 }
